Clean up a little
