+++ src/pages/schedule_old/components/Map/map.module.scss
