+++ src/pages/schedule_old/components/ScheduleEntry.tsx
