+++ src/pages/schedule_old/components/Timer.tsx
